requirements.txt
requests
python-slugify